requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dateutils==0.6.12
Django==3.2.3
django-environ==0.4.5
environ==1.0
idna==2.10
nsetools==1.0.11
psycopg2==2.8.6
python-dateutil==2.8.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
